Import of 'iris.analysis' provides 'iris.analysis.interpolate' again. (#2121)
